+++ towerpy/attc/attc_zhzdr.py
@@ -2,13 +2,17 @@
 
 from pathlib import Path
 import ctypes as ctp
+import copy
 import platform
-import time
 import numpy as np
 import numpy.ctypeslib as npct
 from scipy import optimize
 from ..base import TowerpyError
 from ..utils.radutilities import find_nearest
+from ..utils.radutilities import rolling_window
+from ..utils.radutilities import maf_radial
+from ..utils.radutilities import interp_nan
+from ..utils.radutilities import fillnan1d
 from ..datavis import rad_display
 from ..ml.mlyr import MeltingLayer
 
@@ -41,11 +45,13 @@ class AttenuationCorrection:
         self.site_name = radobj.site_name
 
     def zh_correction(self, rad_georef, rad_params, attvars, cclass, mlyr=None,
+                      phidp_prepro=False,
+                      phidp_prepro_args={'mov_avrgf_len': (1, 3), 't_spdp': 10,
+                                         'minthr_pdp0': -5, 'rhohv_min': 0.90},
                       attc_method='ABRI', pdp_pxavr_rng=7, pdp_pxavr_azm=1,
-                      pdp_dmin=20, coeff_a=[1e-5, 9e-5, 3e-5],
-                      coeff_b=[0.65, 0.85, 0.78],
-                      coeff_alpha=[0.020, 0.1, 0.073], niter=500,
-                      plot_method=False):
+                      pdp_dmin=20, coeff_alpha=[0.020, 0.1, 0.073],
+                      coeff_a=[1e-5, 9e-5, 3e-5], coeff_b=[0.65, 0.85, 0.78],
+                      niter=500, plot_method=False):
         r"""
         Calculate the attenuation of :math:`Z_{H}`.
 
@@ -69,6 +75,34 @@ class AttenuationCorrection:
             corresponding to each azimuth angle of the scan. If None, the
             function is applied to the whole PPI scan, assuming a ml_thickness
             of 0.5 km.
+        phidp_prepro : bool, optional
+            Process :math:`\Phi_{DP}` to mitigate the influence of spurious
+            values within the signal phase. :math:`\Phi_{DP}` must have been
+            previously unfolded and offset (:math:`\Phi_{DP}(r0)`) corrected.
+            The default is False.
+        phidp_prepro_args : dict, optional
+            Parameters used for the preprocessing of :math:`\Phi_{DP}`.
+            The default are:
+
+                mov_avrgf_len : 2-element tuple or list, optional
+                    Window size used to smooth :math:`\Phi_{DP}` by applying a
+                    moving average window. The default is (1, 3). It is
+                    recommended to average :math:`\Phi_{DP}` along the range,
+                    i.e. keep the window size in a (1, n) size.
+
+                t_spdp: int or float, optional
+                    Threshold used to discard bins with standard deviations of
+                    :math:`\Phi_{DP}` greater than the selected value.
+                    The default is 10 deg.
+
+                minthr_pdp0: int or float, optional
+                    Tolerance for the true value of :math:`\Phi_{DP}(r0)`.
+                    Values below this threshold are removed.
+                    The default is -5 deg.
+
+                rhohv_min: float, optional
+                    Threshold used to discard bins related to nonmeteorological
+                    signals. The default is 0.90
         attc_method : str
             Attenuation correction algorithm to be used. The default is 'ABRI':
 
@@ -95,6 +129,11 @@ class AttenuationCorrection:
         pdp_dmin : float
             Minimum total :math:`\Phi_{DP}` expected in a ray to perform
             attenuation correction (at least 20-30 degrees). The default is 20.
+        coeff_alpha : 3-element tuple or list, optional
+            [Min, max, fixed value] of coeff :math:`\alpha`. These bounds are
+            used to find the optimum value of :math:`\alpha` from
+            :math:`A_H = \alpha K_{DP}`. Default values are
+            [0.020, 0.1, 0.073], derived for C-band.
         coeff_a : 3-element tuple or list, optional
             [Min, max, fixed value] of coeff :math:`a`. These bounds are
             used to find the optimum value of :math:`a` from
@@ -105,11 +144,6 @@ class AttenuationCorrection:
             to find the optimum value of :math:`b` from
             :math:`A_H = a Z_{H}^b`. Default values are [0.65, 0.85, 0.78],
             derived for C-band.
-        coeff_alpha : 3-element tuple or list, optional
-            [Min, max, fixed value] of coeff :math:`\alpha`. These bounds are
-            used to find the optimum value of :math:`\alpha` from
-            :math:`A_H = \alpha K_{DP}`. Default values are
-            [0.020, 0.1, 0.073], derived for C-band.
         niter : int
             Number of iterations to find the optimised values of
             the coeffs :math:`a, b, \alpha`. The default is 500.
@@ -149,7 +183,120 @@ class AttenuationCorrection:
             Geoscience and Remote Sensing, vol. 50, no. 12, pp. 5061-5071,
             Dec. 2012. https://doi.org/10.1109/TGRS.2012.2195228
         """
-        tic = time.time()
+        if phidp_prepro:
+            phidp_prepro_argso = {'mov_avrgf_len': (1, 3), 't_spdp': 10,
+                                  'minthr_pdp0': -5, 'rhohv_min': 0.9}
+            phidp_prepro_argso = phidp_prepro_argso | phidp_prepro_args
+            mov_avrgf_len = phidp_prepro_argso['mov_avrgf_len']
+            minthr_pdp0 = phidp_prepro_argso['minthr_pdp0']
+            rhohv_min = phidp_prepro_argso['rhohv_min']
+            t_spdp = phidp_prepro_argso['t_spdp']
+            nrays = rad_params['nrays']
+            ngates = rad_params['ngates']
+            attvars = copy.copy(attvars)
+
+            if (mov_avrgf_len[1] % 2) == 0:
+                print('Choose an odd number to apply the '
+                      + 'moving average filter')
+            phidp_O = {k: np.ones_like(attvars[k]) * attvars[k]
+                       for k in list(attvars) if k.startswith('Phi')}
+            # Removes low-noisy values of PhiDP below the given PhiDP(0)
+            phidp_O['PhiDP [deg]'][phidp_O['PhiDP [deg]']
+                                   < minthr_pdp0] = minthr_pdp0
+            # Filter isolated values
+            phidp_pad = np.pad(phidp_O['PhiDP [deg]'],
+                               ((0, 0), (mov_avrgf_len[1]//2,
+                                         mov_avrgf_len[1]//2)),
+                               mode='constant', constant_values=(np.nan))
+            phidp_dspk = np.array(
+                [[np.nan if ~np.isnan(vbin)
+                  and (np.isnan(phidp_pad[nr][nbin-1])
+                       and np.isnan(phidp_pad[nr][nbin+1]))
+                  else 1 for nbin, vbin in enumerate(phidp_pad[nr])
+                  if nbin != 0 and nbin != phidp_pad.shape[1]-1]
+                 for nr in range(phidp_pad.shape[0])])
+            # Filter using rhoHV
+            phidp_dspk[attvars['rhoHV [-]'] < rhohv_min] = np.nan
+            # Computes sPhiDP for each ray
+            phidp_dspk_rhv = phidp_O['PhiDP [deg]'] * phidp_dspk
+            phidp_s = np.nanstd(rolling_window(
+                phidp_dspk_rhv, mov_avrgf_len), axis=-1, ddof=1)
+            phidp_pad = np.pad(phidp_s, ((0, 0), (mov_avrgf_len[1]//2,
+                                                  mov_avrgf_len[1]//2)),
+                               mode='constant', constant_values=(np.nan))
+            # Filter values with std values greater than std threshold
+            phidp_sfnv = np.array(
+                [[np.nan if vbin >= t_spdp
+                  and (phidp_pad[nr][nbin-1] >= t_spdp
+                       or phidp_pad[nr][nbin+1] >= t_spdp)
+                  else 1 for nbin, vbin in enumerate(phidp_pad[nr])
+                  if nbin != 0 and nbin != phidp_pad.shape[1]-1]
+                 for nr in range(phidp_pad.shape[0])])
+            # Filter isolated values
+            phidp_sfnv2 = np.array(
+                [[np.nan if ~np.isnan(vbin)
+                    and (np.isnan(phidp_pad[nr][nbin-1])
+                         or np.isnan(phidp_pad[nr][nbin+1]))
+                  else 1 for nbin, vbin in enumerate(phidp_pad[nr])
+                  if nbin != 0 and nbin != phidp_pad.shape[1]-1]
+                 for nr in range(phidp_pad.shape[0])])
+            phidp_sfnv = phidp_sfnv*phidp_sfnv2
+            phidp_f = phidp_dspk_rhv * phidp_sfnv
+            # Filter isolated values
+            phidp_pad = np.pad(phidp_f, ((0, 0), (mov_avrgf_len[1]//2,
+                                                  mov_avrgf_len[1]//2)),
+                               mode='constant', constant_values=(np.nan))
+            phidp_f2 = np.array(
+                [[np.nan if ~np.isnan(vbin)
+                  and (np.isnan(phidp_pad[nr][nbin-1])
+                       or np.isnan(phidp_pad[nr][nbin+1]))
+                  else 1 for nbin, vbin in enumerate(phidp_pad[nr])
+                  if nbin != 0 and nbin != phidp_pad.shape[1]-1]
+                 for nr in range(phidp_pad.shape[0])])
+            phidp_f = phidp_f * phidp_f2
+            # Computes and initial PhiDP(0)
+            phidp0 = np.array([[np.nanmedian(
+                nr[np.isfinite(nr)][:mov_avrgf_len[1]+1])
+                if ~np.isnan(nr).all() else 0
+                for nr in phidp_f]]).transpose()
+            phidp0t = np.tile(phidp0, (1, mov_avrgf_len[1] + 1))
+            phidp_f[:, : mov_avrgf_len[1] + 1] = phidp0t
+            # Add phidp0
+            phidp_f = np.array([nr - phidp0[cr]
+                                for cr, nr in enumerate(phidp_f)])
+            # Computes a MAV
+            phidp_m = maf_radial(
+                {'PhiDPi [deg]': phidp_f}, maf_len=mov_avrgf_len[1],
+                maf_ignorenan=True, maf_extendvalid=False)['PhiDPi [deg]']
+            # Filter isolated values
+            phidp_pad = np.pad(phidp_m, ((0, 0), (mov_avrgf_len[1]//2,
+                                                  mov_avrgf_len[1]//2)),
+                               mode='constant', constant_values=(np.nan))
+            phidp_f2 = np.array(
+                [[np.nan if ~np.isnan(vbin)
+                  and (np.isnan(phidp_pad[nr][nbin-1])
+                       or np.isnan(phidp_pad[nr][nbin+1]))
+                  else 1 for nbin, vbin in enumerate(phidp_pad[nr])
+                  if nbin != 0 and nbin != phidp_pad.shape[1]-1]
+                 for nr in range(phidp_pad.shape[0])])
+            phidp_m = phidp_m * phidp_f2
+            # Interpolation and final filtering
+            itprng = np.array(range(ngates))
+            phidp_i = {'PhiDPi [deg]': np.array(
+                [interp_nan(itprng, nr, nan_type='nan') if ~np.isnan(nr).all()
+                 else nr for nr in phidp_m])}
+            phidp_i['PhiDPi [deg]'][:, : mov_avrgf_len[1] + 1] = phidp_f[:, : mov_avrgf_len[1] + 1]
+            phidp_i = {'PhiDPi [deg]': np.array(
+                [fillnan1d(i) for i in phidp_i['PhiDPi [deg]']])}
+            # Apply a moving average filter to the whole PPI
+            phidp_maf = maf_radial(
+                phidp_i, maf_len=mov_avrgf_len[1], maf_ignorenan=False,
+                maf_extendvalid=True)
+            # Filter values using ZH
+            phidp_maf['PhiDPi [deg]'][np.isnan(
+                attvars['ZH [dBZ]'])] = np.nan
+
+            attvars['PhiDP [deg]'] = phidp_maf['PhiDPi [deg]']
 
         array1d = npct.ndpointer(dtype=np.double, ndim=1, flags='CONTIGUOUS')
         array2d = npct.ndpointer(dtype=np.double, ndim=2, flags='CONTIGUOUS')
@@ -174,9 +321,6 @@ class AttenuationCorrection:
             mlyr = MeltingLayer(self)
             mlyr.ml_top = 5
             mlyr.ml_thickness = 0.5
-            mlyr.ml_bottom = mlyr.ml_top - mlyr.ml_thickness
-        else:
-            mlyr.ml_bottom = mlyr.ml_top - mlyr.ml_thickness
 
         if isinstance(mlyr.ml_top, (int, float)):
             mlgrid = np.zeros_like(
@@ -248,18 +392,20 @@ class AttenuationCorrection:
             values[cclass != 0] = np.nan
 
         self.vars = attcorr
-        toc = time.time()
-        print(r'Z_H attenuation correction running time: '
-              f'{toc-tic:.3f} sec.')
+
         if plot_method:
             rad_display.plot_zhattcorr(rad_georef, rad_params, attvars,
-                                       attcorr)
+                                       attcorr, mlyr=mlyr,
+                                       vars_bounds={'alpha [-]':
+                                                    [0,
+                                                     coeff_alpha[1], 11]})
 
     def zdr_correction(self, rad_georef, rad_params, attvars, attcorr_vars,
-                       cclass, mlyr=None, coeff_beta=[0.002, 0.04, 0.02],
-                       rhv_thld=0.98, mov_avrgf_len=5, minbins=10,
-                       p2avrf=3, beta_alpha_ratio=0.265, method='linear',
-                       rparams=None, descr=False, plot_method=False):
+                       cclass, mlyr=None, attc_method='BRI',
+                       coeff_beta=[0.002, 0.07, 0.04],  beta_alpha_ratio=0.265,
+                       rhv_thld=0.985, mov_avrgf_len=9, minbins=10, p2avrf=5,
+                       zh_zdr_model='linear', rparams=None, descr=False,
+                       plot_method=False):
         r"""
         Calculate the attenuation of :math:`Z_{DR}`.
 
@@ -286,35 +432,55 @@ class AttenuationCorrection:
             corresponding to each azimuth angle of the scan. If None, the
             function is applied to the whole PPI scan, assuming a ml_thickness
             of 0.5 km.
+        attc_method : str
+            Attenuation correction algorithm to be used. The default is 'BRI':
+
+                [ABRI] = Bringi (optimised beta parameter).
+
+                [BRI] = Bringi (constant beta/alpha ratio).
+        coeff_beta : 3-element tuple or list, optional
+            [Min, max, fixed value] of coeff :math:`\beta`. These bounds are
+            used to find the optimum value of :math:`\beta`.
+            Default values are [0.002, 0.04, 0.02], derived for C-band.
+        beta_alpha_ratio : float, opt
+            Quotient between :math:`\alpha` and :math:`\beta` parameters from
+            :math:`A_{DP} = ( \beta / \alpha )A_{H}`. The default is 0.265 for
+            C-band.
         rhv_thld : float
             Minimum value of :math:`\rho_{HV}` expected in the rain medium.
             The default is 0.98.
-        minbins : int
-            Minimum number of bins related to the length of each rain cell
-            along the beam. The default is 10.
         mov_avrgf_len : int
             Odd number used to apply a moving average filter to each beam and
             smooth the signal. The default is 5.
+        minbins : int
+            Minimum number of bins related to the length of each rain cell
+            along the beam. The default is 10.
         p2avrf : int
             Number of bins to average on the far side of the rain cell.
             The default is 3.
-        beta_alpha_ratio : float, opt
-            Quotient between :math:`\alpha` and :math:`\beta` parameters from
-            :math:`A_{DP} = ( \beta / \alpha )A_{H}`
-        method : str
+        zh_zdr_model : str
             Method used to compute the ZH-ZDR relationship. The "linear" model
             uses the relation provided in [2]_ and [4]_, whereas the "exp"
             model uses the relation proposed by [3]_.
-        params: dict
-            Additional parameters describing the ZH-ZDR relationship.
+        rparams: dict
+            Additional parameters describing the ZH-ZDR relationship:
+                For the linear model: ZH_lower_lim: 20 dBZ, ZH_upper_lim:
+                45 dBZ, coeff_a: 0.048, coeff_b: 0.774, zdr_max: 1.4
 
-            .. math::
-                Z_{DR} = \Biggl\{ 0 \rightarrow Z_H(r_m)<=20 dBZ \\
-                         0.048*Z_H-0.774 \rightarrow 20<Z_H(r_m)<=45 dBZ \\
-                         1.4 \rightarrow Z_H(r_m)>45 dBZ \Biggl\}
+                .. math::
+                    \overline{Z}_{DR} =
+                    \Biggl\{ 0 \rightarrow  \overline{Z_H}(r_m)<=Z_H(lowerlim) \\
+                            a*Z_H-b \rightarrow Z_H(lowerlim)<Z_H(r_m)<=Z_H(upperlim) \\
+                                Z_{DR}(max) \rightarrow Z_H(r_m)>Z_H(upperlim) \Biggl\}
+
+                For the exp model: coeff_a: 0.00012, coeff_b: 2.5515
+
+                .. math::
+                    \overline{Z}_{DR} =
+                    \Biggl\{ a*Z_H^{b} \Biggl\}
         descr : bool
             Controls if the statistics of the calculations are returned.
-            The default is True.
+            The default is False.
         plot_method : Bool, optional
             Plot the ZDR attenuation correction method. The default is False.
 
@@ -326,14 +492,15 @@ class AttenuationCorrection:
             ADP [dB/km]':
                 Specific differential attenuation.
             beta:
-               parameter :math:`\beta` optimised for a given beam.
+               parameter :math:`\beta` optimised for each beam.
 
         Notes
         -----
         1. The attenuation is computed up to a user-defined melting level
         height.
 
-        2. Based on the method described in [1]_
+        2. The ZDR attenuation correction method assumes that ZH has first been
+        corrected for attenuation, e.g., using the methods described in [1]_
 
         References
         ----------
@@ -361,26 +528,16 @@ class AttenuationCorrection:
             Technology, 22(11), 1621-1632. https://doi.org/10.1175/JTECH1803.1
 
         """
-        tic = time.time()
-        params = {'ZH-ZDR relation': method, 'ZH_lower_lim': 20,
-                  'ZH_upper_lim': 45, 'model': 'a1*ZH-b1', 'zdr_max': 1.4,
-                  'a1': 0.048, 'b1': 0.774}
-        if method == 'exp':
-            params['ZH-ZDR relation'] = method
-            params['model'] = 'a1*ZH^b1'
-            params['a1'] = 0.00012
-            params['b1'] = 2.5515
+        params = {'ZH-ZDR model': zh_zdr_model, 'ZH_lower_lim': 20,
+                  'ZH_upper_lim': 45, 'model': 'coeff_a*ZH-coeff_b',
+                  'zdr_max': 1.4, 'coeff_a': 0.048, 'coeff_b': 0.774}
+        if zh_zdr_model == 'exp':
+            params['ZH-ZDR model'] = zh_zdr_model
+            params['model'] = 'coeff_a*ZH^coeff_b'
+            params['coeff_a'] = 0.00012
+            params['coeff_b'] = 2.5515
         if rparams is not None:
             params.update
-        # if mlyr is None:
-        #     mlyr = MeltingLayer(self)
-        #     mlvl = 5
-        #     mlyr_thickness = 0.5
-        #     # mlyr_bottom = mlvl - mlyr_thickness
-        # else:
-        #     mlvl = mlyr.ml_top
-        #     mlyr_thickness = mlyr.ml_thickness
-        #     mlyr.ml_bottom = mlvl - mlyr_thickness
         if mlyr is None:
             mlyr = MeltingLayer(self)
             mlyr.ml_top = 5
@@ -389,13 +546,6 @@ class AttenuationCorrection:
         else:
             mlyr.ml_bottom = mlyr.ml_top - mlyr.ml_thickness
 
-        if isinstance(mlyr.ml_top, (int, float)):
-            mlt_idx = [find_nearest(nbh, mlyr.ml_top)
-                       for nbh in rad_georef['beam_height [km]']]
-        elif isinstance(mlyr.ml_top, (np.ndarray, list, tuple)):
-            mlt_idx = [find_nearest(nbh, mlyr.ml_top[cnt])
-                       for cnt, nbh in
-                       enumerate(rad_georef['beam_height [km]'])]
         if isinstance(mlyr.ml_bottom, (int, float)):
             mlb_idx = [find_nearest(nbh, mlyr.ml_bottom)
                        for nbh in rad_georef['beam_height [km]']]
@@ -403,20 +553,9 @@ class AttenuationCorrection:
             mlb_idx = [find_nearest(nbh, mlyr.ml_bottom[cnt])
                        for cnt, nbh in
                        enumerate(rad_georef['beam_height [km]'])]
-        # mlt_idxx = np.array([rad_georef['grid_rectx'][cnt, ix]
-        #                      for cnt, ix in enumerate(mlt_idx)])
-        # mlt_idxy = np.array([rad_georef['grid_recty'][cnt, ix]
-        #                      for cnt, ix in enumerate(mlt_idx)])
-        # mlb_idxx = np.array([rad_georef['grid_rectx'][cnt, ix]
-        #                      for cnt, ix in enumerate(mlb_idx)])
-        # mlb_idxy = np.array([rad_georef['grid_recty'][cnt, ix]
-        #                      for cnt, ix in enumerate(mlb_idx)])
-        
-        
+
         nrays = len(rad_georef['azim [rad]'])
-        # TODO fix this so nonisotropic ML are an inpyt.
-        # idxlimfl = [find_nearest(rad_georef['beam_height [km]'][i, :],
-        #             mlvl-mlyr_thickness) for i in range(nrays)]
+
         m = np.zeros_like(attcorr_vars['ZH [dBZ]'])
         for n, rows in enumerate(m):
             rows[mlb_idx[n]:] = np.nan
@@ -436,9 +575,12 @@ class AttenuationCorrection:
         betaof = []
         zdrstat = []
 
-        alphacopy = (np.zeros_like(attcorr_vars['alpha [-]']
-                                   + attcorr_vars['alpha [-]']))
-
+        alphacopy = (np.zeros_like(attcorr_vars['alpha [-]'])
+                     + attcorr_vars['alpha [-]'])
+        if attc_method == 'ABRI':
+            bracket = coeff_beta[:2]
+        elif attc_method == 'BRI':
+            bracket = []
         for i in range(nrays):
             if i in idxzdrattcorr:
                 idmx = np.nancumsum(alphacopy[i]).argmax()
@@ -474,7 +616,7 @@ class AttenuationCorrection:
                         [np.array(range(int(i[0]), int(i[-1])+1))
                          for i in rcells_raw])
                     rcells_dm2 = np.split(
-                        rcells_dm1, np.where(np.diff(rcells_dm1) > 2)[0] + 1)
+                        rcells_dm1, np.where(np.diff(rcells_dm1) > 1)[0] + 1)
                     rcells_crc = [np.array([i[0], i[-1]]).reshape((2, 1))
                                   for i in rcells_dm2]
                     rcells_crc_len = np.concatenate(
@@ -491,18 +633,18 @@ class AttenuationCorrection:
                         zhcrf = np.nanmean(zh_mvavfl[idxrf - p2avrf+1:idxrf+1])
                         zdrmrf = np.nanmean(
                             zdr_mvavfl[idxrf - p2avrf+1:idxrf+1])
-                        if method == 'linear':
+                        if params['ZH-ZDR model'] == 'linear':
                             if zhcrf <= params['ZH_lower_lim']:
                                 zdrerf = 0
                             elif (params['ZH_lower_lim']
                                   < zhcrf <= params['ZH_upper_lim']):
-                                zdrerf = params['a1'] * zhcrf - params['b1']
+                                zdrerf = params['coeff_a'] * zhcrf - params['coeff_b']
                             elif zhcrf > params['ZH_upper_lim']:
                                 zdrerf = params['zdr_max']
                             else:
                                 zdrerf = np.nan
-                        elif method == 'exp':
-                            zdrerf = params['a1']*zhcrf**params['b1']
+                        elif params['ZH-ZDR model'] == 'exp':
+                            zdrerf = params['coeff_a']*zhcrf**params['coeff_b']
                         else:
                             raise TowerpyError('Please check the method '
                                                'selected for estimating the '
@@ -528,26 +670,25 @@ class AttenuationCorrection:
                                              * np.nanmean(
                                                  attcorrmask['PIA [dB]'][i, :]
                                                  [idxrf-p2avrf+1:idxrf+1])))
-                                        - zdrerf,
-                                        bracket=[coeff_beta[:2]],
+                                        - zdrerf, bracket=bracket,
                                         x0=(np.nanmin(alphacopy[i])
                                             * beta_alpha_ratio),
                                         method='brentq')
                                     betao = sl1.root
                                 else:
                                     betao = betai
-                                if betao <= 0:
+                                if betao <= coeff_beta[0]:
                                     betao = coeff_beta[0]
                                 if betao >= coeff_beta[1]:
                                     betao = coeff_beta[1]
+                                betaopt = (np.zeros_like(
+                                    attcorr_vars['alpha [-]'][i, :]) + betao)
                                 zdrcr = ((attvars['ZDR [dB]'][i, :])
                                          + ((betao/alphacopy[i, :])
                                             * attcorr_vars['PIA [dB]'][i, :]))
                                 adpi = ((betao
                                          / attcorr_vars['alpha [-]'][i, :])
                                         * attcorr_vars['AH [dB/km]'][i, :])
-                                betaopt = (np.zeros_like(
-                                    attcorr_vars['alpha [-]'][i, :]) + betao)
                                 statzdr = f'{i}: beta coeff optimised 1 iter'
                             except ValueError:
                                 idxrs = int(raincells[0][0])
@@ -556,19 +697,19 @@ class AttenuationCorrection:
                                                              idxrf+1])
                                 zdrmrf = np.nanmean(zdr_mvavfl[idxrf-p2avrf+1:
                                                                idxrf+1])
-                                if method == 'linear':
+                                if params['ZH-ZDR model'] == 'linear':
                                     if zhcrf <= params['ZH_lower_lim']:
                                         zdrerf = 0
                                     elif (params['ZH_lower_lim']
                                           < zhcrf <= params['ZH_upper_lim']):
-                                        zdrerf = (params['a1']
-                                                  * zhcrf - params['b1'])
+                                        zdrerf = (params['coeff_a']
+                                                  * zhcrf - params['coeff_b'])
                                     elif zhcrf > params['ZH_upper_lim']:
                                         zdrerf = params['zdr_max']
                                     else:
                                         zdrerf = np.nan
-                                elif method == 'exp':
-                                    zdrerf = params['a1']*zhcrf**params['b1']
+                                elif params['ZH-ZDR model'] == 'exp':
+                                    zdrerf = params['coeff_a']*zhcrf**params['coeff_b']
                                 else:
                                     raise
                                     TowerpyError('Please check the method '
@@ -591,14 +732,14 @@ class AttenuationCorrection:
                                             betaif: ((zdrmrf) + (betaif) *
                                                      ((1 / attcorrmask['alpha [-]'][i, :][idxrf])
                                                       * np.nanmean(attcorrmask['PIA [dB]'][i, :][idxrf - p2avrf+1:idxrf+1]))) - zdrerf,
-                                            bracket=[coeff_beta[:2]],
+                                            bracket=bracket,
                                             x0=(np.nanmin(alphacopy[i])
                                                 * beta_alpha_ratio),
                                             method='brentq')
                                         betao = sl1.root
                                     else:
                                         betao = betai
-                                    if betao <= 0:
+                                    if betao <= coeff_beta[0]:
                                         betao = coeff_beta[0]
                                     if betao >= coeff_beta[1]:
                                         betao = coeff_beta[1]
@@ -684,9 +825,6 @@ class AttenuationCorrection:
         self.res_zdrcorrection = zdr_calc
         self.vars |= attcorr1
 
-        toc = time.time()
-        print(r'Z_{DR} attenuation correction running time: '
-              f'{toc-tic:.3f} sec.')
         if plot_method:
             rad_display.plot_zdrattcorr(rad_georef, rad_params, attvars,
-                                        attcorr1)
+                                        attcorr1, mlyr=mlyr)